moved bookmark components in their folder
