+++ src/components/bookmark/AddBookmarkModal.tsx
@@ -5,15 +5,15 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog';
-import { Label } from './ui/label';
-import { Input } from './ui/input';
-import { Button } from './ui/button';
-import { Textarea } from './ui/textarea';
+import { Label } from '../ui/label';
+import { Input } from '../ui/input';
+import { Button } from '../ui/button';
+import { Textarea } from '../ui/textarea';
 import { formatTime } from '@/lib/utils';
 import { useAction } from '@/hooks/useAction';
 import { toast } from 'sonner';
 import { createBookmark } from '@/actions/bookmark';
-import { FormErrors } from './FormError';
+import { FormErrors } from '../FormError';
 import { useParams } from 'next/navigation';
 import { FormEvent } from 'react';
